simplify tui: extract sub-model update helpers, move tab logic to its own file
- centralise the chatPage/editor/dialog/completions Update boilerplate

  in dispatch.go (forwardX returns (m, cmd); updateXCmd returns just cmd

  for batching)

- move tab restore/persist/lookup logic from tui.go to tabs.go

- extract copyToClipboard helper to deduplicate clipboard handlers

- replace isErrTitleGenerating string comparison with errors.Is

- inline persistSplitDiffView helper for the userconfig save
