+++ pkg/tui/tabs.go
@@ -1,0 +1,171 @@
+package tui
+
+import (
+	"context"
+	"log/slog"
+
+	"github.com/docker/docker-agent/pkg/app"
+	"github.com/docker/docker-agent/pkg/tui/service/supervisor"
+	"github.com/docker/docker-agent/pkg/tui/service/tuistate"
+	"github.com/docker/docker-agent/pkg/userconfig"
+)
+
+// This file contains the persistence and lookup logic for the multi-tab
+// model. The TUI uses two parallel tab identifiers:
+//
+//   - "runtime tab IDs": ephemeral, generated by the supervisor for the live
+//     session of an open tab.
+//   - "persisted session IDs": stable session-store IDs that survive
+//     restarts and are stored in tuistate.
+//
+// pendingRestores bridges the two until a tab is lazily loaded on first
+// switch.
+
+// persistActiveTab writes the active tab ID to the tuistate store.
+func (m *appModel) persistActiveTab(persistedID string) {
+	if m.tuiStore == nil {
+		return
+	}
+	if err := m.tuiStore.SetActiveTab(context.Background(), persistedID); err != nil {
+		slog.Warn("Failed to set active tab", "error", err)
+	}
+}
+
+// persistFreshTab clears the tab store and writes a single initial tab.
+func (m *appModel) persistFreshTab(ctx context.Context, sessionID, workingDir string) {
+	if m.tuiStore == nil {
+		return
+	}
+	if err := m.tuiStore.ClearTabs(ctx); err != nil {
+		slog.Warn("Failed to clear tabs", "error", err)
+	}
+	if err := m.tuiStore.AddTab(ctx, sessionID, workingDir); err != nil {
+		slog.Warn("Failed to persist initial tab", "error", err)
+	}
+	if err := m.tuiStore.SetActiveTab(ctx, sessionID); err != nil {
+		slog.Warn("Failed to set active tab", "error", err)
+	}
+}
+
+// restoreTabs restores previously persisted tabs (if enabled) or persists the
+// initial session as the sole tab. The tuistate DB always stores persisted
+// session-store IDs. Runtime tab/routing IDs are ephemeral; the pendingRestores
+// map bridges the two:  pendingRestores[runtimeTabID] = persistedSessionID.
+func (m *appModel) restoreTabs(
+	ctx context.Context,
+	ts *tuistate.Store,
+	sv *supervisor.Supervisor,
+	spawner SessionSpawner,
+	initialApp *app.App,
+	initialTabID, initialWorkingDir string,
+) {
+	if ts == nil {
+		return
+	}
+
+	var savedTabs []tuistate.TabEntry
+	var savedActiveID string
+	if *userconfig.Get().RestoreTabs {
+		savedTabs, savedActiveID, _ = ts.GetTabs(ctx)
+	}
+
+	if len(savedTabs) == 0 {
+		m.persistFreshTab(ctx, initialTabID, initialWorkingDir)
+		return
+	}
+
+	sessionStore := initialApp.SessionStore()
+	restoredFirst := false
+
+	for _, saved := range savedTabs {
+		// Validate the saved session still exists.
+		if sessionStore != nil && saved.SessionID != "" {
+			if _, err := sessionStore.GetSession(ctx, saved.SessionID); err != nil {
+				slog.Warn("Saved session no longer exists, removing stale tab",
+					"session_id", saved.SessionID, "error", err)
+				_ = ts.RemoveTab(ctx, saved.SessionID)
+				continue
+			}
+		}
+
+		// Determine the runtime tab ID to use.
+		var runtimeID string
+		if !restoredFirst {
+			restoredFirst = true
+			runtimeID = initialTabID
+		} else {
+			a, newSess, spawnCleanup, err := spawner(ctx, saved.WorkingDir)
+			if err != nil {
+				slog.Warn("Failed to restore tab", "working_dir", saved.WorkingDir, "error", err)
+				_ = ts.RemoveTab(ctx, saved.SessionID)
+				continue
+			}
+			runtimeID = sv.AddSession(ctx, a, newSess, saved.WorkingDir, spawnCleanup)
+		}
+
+		// Stash persisted session ID for lazy loading on first switch.
+		m.pendingRestores[runtimeID] = saved.SessionID
+		if saved.SidebarCollapsed {
+			m.pendingSidebarCollapsed[runtimeID] = true
+		}
+
+		// If this was the active tab, queue a switch on Init().
+		if saved.SessionID == savedActiveID {
+			if restoredFirst && runtimeID == initialTabID {
+				_ = ts.SetActiveTab(ctx, saved.SessionID)
+			} else {
+				m.pendingActiveTab = runtimeID
+			}
+		}
+
+		// Peek at the session title so the tab bar shows a name before lazy load.
+		if sessionStore != nil && saved.SessionID != "" {
+			if oldSess, err := sessionStore.GetSession(ctx, saved.SessionID); err == nil && oldSess.Title != "" {
+				sv.SetRunnerTitle(runtimeID, oldSess.Title)
+			}
+		}
+	}
+
+	// If all saved tabs were stale, persist the initial session.
+	if !restoredFirst {
+		m.persistFreshTab(ctx, initialTabID, initialWorkingDir)
+	}
+}
+
+// persistedSessionID returns the session-store ID that should be used for
+// tuistate persistence for the given runtime tab ID.
+//
+// If the tab has a pending restore (session not yet lazily loaded), the
+// persisted ID from pendingRestores is returned — this is the original
+// session-store ID that was saved across restarts. Otherwise the live
+// session ID from the app is used.
+func (m *appModel) persistedSessionID(tabID string) string {
+	if persistedID, ok := m.pendingRestores[tabID]; ok {
+		return persistedID
+	}
+	if runner := m.supervisor.GetRunner(tabID); runner != nil {
+		return runner.App.Session().ID
+	}
+	return tabID
+}
+
+// findTabByPersistedID scans all open tabs and returns the runtime tab ID
+// whose persisted session-store ID matches the given ID. Returns "" if not found.
+func (m *appModel) findTabByPersistedID(persistedID string) string {
+	// Check pending restores first (tabs not yet lazily loaded).
+	for tabID, pid := range m.pendingRestores {
+		if pid == persistedID {
+			return tabID
+		}
+	}
+	// Check live sessions.
+	tabs, _ := m.supervisor.GetTabs()
+	for _, tab := range tabs {
+		if runner := m.supervisor.GetRunner(tab.SessionID); runner != nil {
+			if runner.App.Session().ID == persistedID {
+				return tab.SessionID
+			}
+		}
+	}
+	return ""
+}
