+++ pkg/tui/tui.go
@@ -32,7 +32,6 @@ import (
 	"github.com/docker/docker-agent/pkg/tui/components/statusbar"
 	"github.com/docker/docker-agent/pkg/tui/components/tabbar"
 	"github.com/docker/docker-agent/pkg/tui/core"
-	"github.com/docker/docker-agent/pkg/tui/core/layout"
 	"github.com/docker/docker-agent/pkg/tui/dialog"
 	"github.com/docker/docker-agent/pkg/tui/internal/editorname"
 	"github.com/docker/docker-agent/pkg/tui/messages"
@@ -364,10 +363,8 @@ func (m *appModel) reapplyKeyboardEnhancements() {
 	if m.keyboardEnhancements == nil {
 		return
 	}
-	updated, _ := m.chatPage.Update(*m.keyboardEnhancements)
-	m.chatPage = updated.(chat.Page)
-	editorModel, _ := m.editor.Update(*m.keyboardEnhancements)
-	m.editor = editorModel.(editor.Editor)
+	_ = m.updateChatCmd(*m.keyboardEnhancements)
+	_ = m.updateEditorCmd(*m.keyboardEnhancements)
 }
 
 func (m *appModel) commandCategories() []commands.Category {
@@ -425,117 +422,6 @@ func (m *appModel) initAndFocusComponents() tea.Cmd {
 		m.editor.Focus(),
 		m.resizeAll(),
 	)
-}
-
-// persistActiveTab writes the active tab ID to the tuistate store.
-func (m *appModel) persistActiveTab(persistedID string) {
-	if m.tuiStore == nil {
-		return
-	}
-	if err := m.tuiStore.SetActiveTab(context.Background(), persistedID); err != nil {
-		slog.Warn("Failed to set active tab", "error", err)
-	}
-}
-
-// persistFreshTab clears the tab store and writes a single initial tab.
-func (m *appModel) persistFreshTab(ctx context.Context, sessionID, workingDir string) {
-	if m.tuiStore == nil {
-		return
-	}
-	if err := m.tuiStore.ClearTabs(ctx); err != nil {
-		slog.Warn("Failed to clear tabs", "error", err)
-	}
-	if err := m.tuiStore.AddTab(ctx, sessionID, workingDir); err != nil {
-		slog.Warn("Failed to persist initial tab", "error", err)
-	}
-	if err := m.tuiStore.SetActiveTab(ctx, sessionID); err != nil {
-		slog.Warn("Failed to set active tab", "error", err)
-	}
-}
-
-// restoreTabs restores previously persisted tabs (if enabled) or persists the
-// initial session as the sole tab. The tuistate DB always stores persisted
-// session-store IDs. Runtime tab/routing IDs are ephemeral; the pendingRestores
-// map bridges the two:  pendingRestores[runtimeTabID] = persistedSessionID
-func (m *appModel) restoreTabs(
-	ctx context.Context,
-	ts *tuistate.Store,
-	sv *supervisor.Supervisor,
-	spawner SessionSpawner,
-	initialApp *app.App,
-	initialTabID, initialWorkingDir string,
-) {
-	if ts == nil {
-		return
-	}
-
-	var savedTabs []tuistate.TabEntry
-	var savedActiveID string
-	if *userconfig.Get().RestoreTabs {
-		savedTabs, savedActiveID, _ = ts.GetTabs(ctx)
-	}
-
-	if len(savedTabs) == 0 {
-		m.persistFreshTab(ctx, initialTabID, initialWorkingDir)
-		return
-	}
-
-	sessionStore := initialApp.SessionStore()
-	restoredFirst := false
-
-	for _, saved := range savedTabs {
-		// Validate the saved session still exists.
-		if sessionStore != nil && saved.SessionID != "" {
-			if _, err := sessionStore.GetSession(ctx, saved.SessionID); err != nil {
-				slog.Warn("Saved session no longer exists, removing stale tab",
-					"session_id", saved.SessionID, "error", err)
-				_ = ts.RemoveTab(ctx, saved.SessionID)
-				continue
-			}
-		}
-
-		// Determine the runtime tab ID to use.
-		var runtimeID string
-		if !restoredFirst {
-			restoredFirst = true
-			runtimeID = initialTabID
-		} else {
-			a, newSess, spawnCleanup, err := spawner(ctx, saved.WorkingDir)
-			if err != nil {
-				slog.Warn("Failed to restore tab", "working_dir", saved.WorkingDir, "error", err)
-				_ = ts.RemoveTab(ctx, saved.SessionID)
-				continue
-			}
-			runtimeID = sv.AddSession(ctx, a, newSess, saved.WorkingDir, spawnCleanup)
-		}
-
-		// Stash persisted session ID for lazy loading on first switch.
-		m.pendingRestores[runtimeID] = saved.SessionID
-		if saved.SidebarCollapsed {
-			m.pendingSidebarCollapsed[runtimeID] = true
-		}
-
-		// If this was the active tab, queue a switch on Init().
-		if saved.SessionID == savedActiveID {
-			if restoredFirst && runtimeID == initialTabID {
-				_ = ts.SetActiveTab(ctx, saved.SessionID)
-			} else {
-				m.pendingActiveTab = runtimeID
-			}
-		}
-
-		// Peek at the session title so the tab bar shows a name before lazy load.
-		if sessionStore != nil && saved.SessionID != "" {
-			if oldSess, err := sessionStore.GetSession(ctx, saved.SessionID); err == nil && oldSess.Title != "" {
-				sv.SetRunnerTitle(runtimeID, oldSess.Title)
-			}
-		}
-	}
-
-	// If all saved tabs were stale, persist the initial session.
-	if !restoredFirst {
-		m.persistFreshTab(ctx, initialTabID, initialWorkingDir)
-	}
 }
 
 // Init initializes the model.
@@ -602,14 +488,10 @@ func (m *appModel) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 		return m.handleRoutedMsg(msg)
 
 	case animation.TickMsg:
-		var cmds []tea.Cmd
-		updated, cmd := m.chatPage.Update(msg)
-		m.chatPage = updated.(chat.Page)
-		cmds = append(cmds, cmd)
+		cmds := []tea.Cmd{m.updateChatCmd(msg)}
 		// Update working spinner
 		if m.chatPage.IsWorking() {
-			var model layout.Model
-			model, cmd = m.workingSpinner.Update(msg)
+			model, cmd := m.workingSpinner.Update(msg)
 			m.workingSpinner = model.(spinner.Spinner)
 			cmds = append(cmds, cmd)
 		}
@@ -733,13 +615,7 @@ func (m *appModel) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 	case tea.KeyboardEnhancementsMsg:
 		m.keyboardEnhancements = &msg
 		m.keyboardEnhancementsSupported = msg.Flags != 0
-		// Forward to content view
-		updated, cmd := m.chatPage.Update(msg)
-		m.chatPage = updated.(chat.Page)
-		// Forward to editor
-		editorModel, editorCmd := m.editor.Update(msg)
-		m.editor = editorModel.(editor.Editor)
-		return m, tea.Batch(cmd, editorCmd)
+		return m, tea.Batch(m.updateChatCmd(msg), m.updateEditorCmd(msg))
 
 	// --- Keyboard input ---
 
@@ -748,21 +624,15 @@ func (m *appModel) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 
 	case tea.PasteMsg:
 		if m.dialogMgr.Open() {
-			u, cmd := m.dialogMgr.Update(msg)
-			m.dialogMgr = u.(dialog.Manager)
-			return m, cmd
+			return m.forwardDialog(msg)
 		}
 		// When inline editing a past message, forward paste to the chat page
 		// so the messages component can insert content into the inline textarea.
 		if m.chatPage.IsInlineEditing() {
-			updated, cmd := m.chatPage.Update(msg)
-			m.chatPage = updated.(chat.Page)
-			return m, cmd
+			return m.forwardChat(msg)
 		}
 		// Forward paste to editor
-		editorModel, cmd := m.editor.Update(msg)
-		m.editor = editorModel.(editor.Editor)
-		return m, cmd
+		return m.forwardEditor(msg)
 
 	// --- Mouse ---
 
@@ -781,9 +651,7 @@ func (m *appModel) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 	// --- Dialog lifecycle ---
 
 	case dialog.OpenDialogMsg, dialog.CloseDialogMsg:
-		u, cmd := m.dialogMgr.Update(msg)
-		m.dialogMgr = u.(dialog.Manager)
-		return m, cmd
+		return m.forwardDialog(msg)
 
 	case dialog.ExitConfirmedMsg:
 		m.cleanupAll()
@@ -828,22 +696,16 @@ func (m *appModel) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 	case *runtime.TeamInfoEvent:
 		m.sessionState.SetAvailableAgents(msg.AvailableAgents)
 		m.sessionState.SetCurrentAgentName(msg.CurrentAgent)
-		updated, cmd := m.chatPage.Update(msg)
-		m.chatPage = updated.(chat.Page)
-		return m, cmd
+		return m.forwardChat(msg)
 
 	case *runtime.AgentInfoEvent:
 		m.sessionState.SetCurrentAgentName(msg.AgentName)
 		m.application.TrackCurrentAgentModel(msg.Model)
-		updated, cmd := m.chatPage.Update(msg)
-		m.chatPage = updated.(chat.Page)
-		return m, cmd
+		return m.forwardChat(msg)
 
 	case *runtime.SessionTitleEvent:
 		m.sessionState.SetSessionTitle(msg.Title)
-		updated, cmd := m.chatPage.Update(msg)
-		m.chatPage = updated.(chat.Page)
-		return m, cmd
+		return m.forwardChat(msg)
 
 	// --- New session (slash command /new) ---
 
@@ -877,9 +739,7 @@ func (m *appModel) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 		if m.history != nil {
 			_ = m.history.Add(msg.Content)
 		}
-		updated, cmd := m.chatPage.Update(msg)
-		m.chatPage = updated.(chat.Page)
-		return m, cmd
+		return m.forwardChat(msg)
 
 	// --- File attachments (routed to editor) ---
 
@@ -921,9 +781,7 @@ func (m *appModel) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 		return m.handleToggleSplitDiff()
 
 	case messages.ClearQueueMsg:
-		updated, cmd := m.chatPage.Update(msg)
-		m.chatPage = updated.(chat.Page)
-		return m, cmd
+		return m.forwardChat(msg)
 
 	case messages.CompactSessionMsg:
 		return m.handleCompactSession(msg.AdditionalPrompt)
@@ -1058,33 +916,16 @@ func (m *appModel) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 			if agentName := event.GetAgentName(); agentName != "" {
 				m.sessionState.SetCurrentAgentName(agentName)
 			}
-			updated, cmd := m.chatPage.Update(msg)
-			m.chatPage = updated.(chat.Page)
-			return m, cmd
+			return m.forwardChat(msg)
 		}
 
-		// Forward to dialog if open
+		// Forward to dialog if open (and to chat in parallel)
 		if m.dialogMgr.Open() {
-			u, cmd := m.dialogMgr.Update(msg)
-			m.dialogMgr = u.(dialog.Manager)
-
-			updated, cmdChatPage := m.chatPage.Update(msg)
-			m.chatPage = updated.(chat.Page)
-
-			return m, tea.Batch(cmd, cmdChatPage)
+			return m, tea.Batch(m.updateDialogCmd(msg), m.updateChatCmd(msg))
 		}
 
-		// Forward to both completion manager and editor
-		updatedComp, cmdCompletions := m.completions.Update(msg)
-		m.completions = updatedComp.(completion.Manager)
-
-		editorModel, cmdEditor := m.editor.Update(msg)
-		m.editor = editorModel.(editor.Editor)
-
-		updated, cmdChatPage := m.chatPage.Update(msg)
-		m.chatPage = updated.(chat.Page)
-
-		return m, tea.Batch(cmdCompletions, cmdEditor, cmdChatPage)
+		// Forward to completion manager, editor, and chat page in parallel
+		return m, tea.Batch(m.updateCompletionsCmd(msg), m.updateEditorCmd(msg), m.updateChatCmd(msg))
 	}
 }
 
@@ -1654,9 +1495,7 @@ func (m *appModel) resizeAll() tea.Cmd {
 	}
 
 	// Full mode: update overlay components
-	u, cmd := m.dialogMgr.Update(tea.WindowSizeMsg{Width: width, Height: height})
-	m.dialogMgr = u.(dialog.Manager)
-	cmds = append(cmds, cmd)
+	cmds = append(cmds, m.updateDialogCmd(tea.WindowSizeMsg{Width: width, Height: height}))
 
 	m.completions.SetEditorBottom(editorHeight + m.tabBar.Height())
 	m.completions.Update(tea.WindowSizeMsg{Width: width, Height: height})
@@ -1822,9 +1661,7 @@ func (m *appModel) handleKeyPress(msg tea.KeyPressMsg) (tea.Model, tea.Cmd) {
 
 	// Dialog gets priority when open
 	if m.dialogMgr.Open() {
-		u, cmd := m.dialogMgr.Update(msg)
-		m.dialogMgr = u.(dialog.Manager)
-		return m, cmd
+		return m.forwardDialog(msg)
 	}
 
 	// Tab bar keys (Ctrl+t, Ctrl+p, Ctrl+n, Ctrl+w) are suppressed during
@@ -1841,20 +1678,10 @@ func (m *appModel) handleKeyPress(msg tea.KeyPressMsg) (tea.Model, tea.Cmd) {
 	// Completion popup gets priority when open
 	if m.completions.Open() {
 		if core.IsNavigationKey(msg) {
-			u, cmd := m.completions.Update(msg)
-			m.completions = u.(completion.Manager)
-			return m, cmd
+			return m.forwardCompletions(msg)
 		}
 		// For all other keys (typing), send to both completion (for filtering) and editor
-		var cmds []tea.Cmd
-		u, completionCmd := m.completions.Update(msg)
-		m.completions = u.(completion.Manager)
-		cmds = append(cmds, completionCmd)
-
-		editorModel, cmd := m.editor.Update(msg)
-		m.editor = editorModel.(editor.Editor)
-		cmds = append(cmds, cmd)
-		return m, tea.Batch(cmds...)
+		return m, tea.Batch(m.updateCompletionsCmd(msg), m.updateEditorCmd(msg))
 	}
 
 	// Global keyboard shortcuts (active even during history search)
@@ -1897,9 +1724,7 @@ func (m *appModel) handleKeyPress(msg tea.KeyPressMsg) (tea.Model, tea.Cmd) {
 
 	// History search is a modal state — capture all remaining keys before normal routing
 	if m.focusedPanel == PanelEditor && m.editor.IsHistorySearchActive() {
-		editorModel, cmd := m.editor.Update(msg)
-		m.editor = editorModel.(editor.Editor)
-		return m, cmd
+		return m.forwardEditor(msg)
 	}
 
 	switch {
@@ -1918,9 +1743,7 @@ func (m *appModel) handleKeyPress(msg tea.KeyPressMsg) (tea.Model, tea.Cmd) {
 		if m.leanMode {
 			return m, nil
 		}
-		updated, cmd := m.chatPage.Update(msg)
-		m.chatPage = updated.(chat.Page)
-		return m, cmd
+		return m.forwardChat(msg)
 
 	// Focus switching: Tab key toggles between content and editor
 	case key.Matches(msg, key.NewBinding(key.WithKeys("tab"))):
@@ -1929,9 +1752,7 @@ func (m *appModel) handleKeyPress(msg tea.KeyPressMsg) (tea.Model, tea.Cmd) {
 	// Esc: cancel stream (works regardless of focus)
 	case key.Matches(msg, key.NewBinding(key.WithKeys("esc"))):
 		// Forward to content view for stream cancellation
-		updated, cmd := m.chatPage.Update(msg)
-		m.chatPage = updated.(chat.Page)
-		return m, cmd
+		return m.forwardChat(msg)
 
 	default:
 		// Handle ctrl+1 through ctrl+9 for quick agent switching
@@ -1943,13 +1764,9 @@ func (m *appModel) handleKeyPress(msg tea.KeyPressMsg) (tea.Model, tea.Cmd) {
 	// Focus-based routing
 	switch m.focusedPanel {
 	case PanelEditor:
-		editorModel, cmd := m.editor.Update(msg)
-		m.editor = editorModel.(editor.Editor)
-		return m, cmd
+		return m.forwardEditor(msg)
 	case PanelContent:
-		updated, cmd := m.chatPage.Update(msg)
-		m.chatPage = updated.(chat.Page)
-		return m, cmd
+		return m.forwardChat(msg)
 	}
 
 	return m, nil
@@ -1994,18 +1811,14 @@ func (m *appModel) handleMouseClick(msg tea.MouseClickMsg) (tea.Model, tea.Cmd) 
 
 	// Dialogs use full-window coordinates (they're positioned over the entire screen)
 	if m.dialogMgr.Open() {
-		u, cmd := m.dialogMgr.Update(msg)
-		m.dialogMgr = u.(dialog.Manager)
-		return m, cmd
+		return m.forwardDialog(msg)
 	}
 
 	region := m.hitTestRegion(msg.Y)
 
 	switch region {
 	case regionContent:
-		updated, cmd := m.chatPage.Update(msg)
-		m.chatPage = updated.(chat.Page)
-		return m, cmd
+		return m.forwardChat(msg)
 
 	case regionResizeHandle:
 		if msg.Button == tea.MouseLeft {
@@ -2034,9 +1847,7 @@ func (m *appModel) handleMouseClick(msg tea.MouseClickMsg) (tea.Model, tea.Cmd) 
 		adjustedMsg := msg
 		adjustedMsg.X = msg.X - styles.AppPadding
 		adjustedMsg.Y = msg.Y - m.editorTop()
-		editorModel, cmd := m.editor.Update(adjustedMsg)
-		m.editor = editorModel.(editor.Editor)
-		return m, tea.Batch(cmd, m.editor.Focus())
+		return m, tea.Batch(m.updateEditorCmd(adjustedMsg), m.editor.Focus())
 
 	case regionStatusBar:
 		if msg.Button == tea.MouseLeft && m.statusBar.ClickedNewTab(msg.X) {
@@ -2063,9 +1874,7 @@ func (m *appModel) handleMouseMotion(msg tea.MouseMotionMsg) (tea.Model, tea.Cmd
 	}
 
 	if m.dialogMgr.Open() {
-		u, cmd := m.dialogMgr.Update(msg)
-		m.dialogMgr = u.(dialog.Manager)
-		return m, cmd
+		return m.forwardDialog(msg)
 	}
 
 	// Update hover state for resize handle
@@ -2073,16 +1882,12 @@ func (m *appModel) handleMouseMotion(msg tea.MouseMotionMsg) (tea.Model, tea.Cmd
 	m.isHoveringHandle = region == regionResizeHandle
 	switch region {
 	case regionContent:
-		updated, cmd := m.chatPage.Update(msg)
-		m.chatPage = updated.(chat.Page)
-		return m, cmd
+		return m.forwardChat(msg)
 	case regionEditor:
 		adjustedMsg := msg
 		adjustedMsg.X = msg.X - styles.AppPadding
 		adjustedMsg.Y = msg.Y - m.editorTop()
-		editorModel, cmd := m.editor.Update(adjustedMsg)
-		m.editor = editorModel.(editor.Editor)
-		return m, cmd
+		return m.forwardEditor(adjustedMsg)
 	}
 
 	return m, nil
@@ -2106,24 +1911,18 @@ func (m *appModel) handleMouseRelease(msg tea.MouseReleaseMsg) (tea.Model, tea.C
 	}
 
 	if m.dialogMgr.Open() {
-		u, cmd := m.dialogMgr.Update(msg)
-		m.dialogMgr = u.(dialog.Manager)
-		return m, cmd
+		return m.forwardDialog(msg)
 	}
 
 	region := m.hitTestRegion(msg.Y)
 	switch region {
 	case regionContent:
-		updated, cmd := m.chatPage.Update(msg)
-		m.chatPage = updated.(chat.Page)
-		return m, cmd
+		return m.forwardChat(msg)
 	case regionEditor:
 		adjustedMsg := msg
 		adjustedMsg.X = msg.X - styles.AppPadding
 		adjustedMsg.Y = msg.Y - m.editorTop()
-		editorModel, cmd := m.editor.Update(adjustedMsg)
-		m.editor = editorModel.(editor.Editor)
-		return m, cmd
+		return m.forwardEditor(adjustedMsg)
 	}
 
 	return m, nil
@@ -2136,17 +1935,13 @@ func (m *appModel) handleWheelCoalesced(msg messages.WheelCoalescedMsg) (tea.Mod
 	}
 
 	if m.dialogMgr.Open() {
-		u, cmd := m.dialogMgr.Update(msg)
-		m.dialogMgr = u.(dialog.Manager)
-		return m, cmd
+		return m.forwardDialog(msg)
 	}
 
 	region := m.hitTestRegion(msg.Y)
 	switch region {
 	case regionContent:
-		updated, cmd := m.chatPage.Update(msg)
-		m.chatPage = updated.(chat.Page)
-		return m, cmd
+		return m.forwardChat(msg)
 	case regionEditor:
 		m.editor.ScrollByWheel(msg.Delta)
 		return m, nil
@@ -2430,44 +2225,6 @@ func (m *appModel) cleanupAll() {
 		slog.Warn("Graceful shutdown timed out, forcing exit")
 		exitFunc(0)
 	}()
-}
-
-// persistedSessionID returns the session-store ID that should be used for
-// tuistate persistence for the given runtime tab ID.
-//
-// If the tab has a pending restore (session not yet lazily loaded), the
-// persisted ID from pendingRestores is returned — this is the original
-// session-store ID that was saved across restarts. Otherwise the live
-// session ID from the app is used.
-func (m *appModel) persistedSessionID(tabID string) string {
-	if persistedID, ok := m.pendingRestores[tabID]; ok {
-		return persistedID
-	}
-	if runner := m.supervisor.GetRunner(tabID); runner != nil {
-		return runner.App.Session().ID
-	}
-	return tabID
-}
-
-// findTabByPersistedID scans all open tabs and returns the runtime tab ID
-// whose persisted session-store ID matches the given ID. Returns "" if not found.
-func (m *appModel) findTabByPersistedID(persistedID string) string {
-	// Check pending restores first (tabs not yet lazily loaded).
-	for tabID, pid := range m.pendingRestores {
-		if pid == persistedID {
-			return tabID
-		}
-	}
-	// Check live sessions.
-	tabs, _ := m.supervisor.GetTabs()
-	for _, tab := range tabs {
-		if runner := m.supervisor.GetRunner(tab.SessionID); runner != nil {
-			if runner.App.Session().ID == persistedID {
-				return tab.SessionID
-			}
-		}
-	}
-	return ""
 }
 
 // openExternalEditor opens the current editor content in an external editor.
